Added some new hamster art.
Added a delay to multiple bubbles popping so that they pop one after the other.

Made it so a line of bubbles will only be added if there are no bubbles in midair. This helps prevents bubbles from being "swallowed" by the board moving down as it gets close to the board.

Set up survival mode to show game time.

Changed Node's CheckCanBeHit function to a coroutine for optimization purposes.