Greatly improved AIAction weight determine code. Increasing performance especially when dealing with opponents and shifting.
Added new forest art assets.

Reworked AttackState to rush the player forward with a static hitbox instead of having the punch attack move forward from the player.

Changed hamster spawns to make the special hamster type not use the random seed (so it can be different per player).

Optimized and improved checking of if the board is stable.
Optimized and improved checking of if a node is relevant to the AI or not.

Adjusted demo mode to skip Game Setup and go straight to the match.

Added new font.

Fixed bug that caused Pause to persist when leaving a match from the pause menu.
Fixed bug that caused player to not consider the cooldown between attacks.