Huge asset dump. Boy animations, hamster stuff, etc.
Big update to player animations to work with the Boy animations.

Setup animations to move held bubble to match animations.
Adjusted throw to happen on specific animation frame.

Replaced hamster sprites in bubbles to hamster faces.

Wrote a shader to display the Shift Meter better to prepare for upcoming assets.