Unity 2018 Update
Updated the game to Unity version 2018.4.2f1. Fixed any issues involved with updating to the new Unity version.

Refactored how board scenes work. Instead of dozens of scenes for every board and game mode, now there is only 1 scene per game mode and the stages are loaded in when the scene starts.

Also uploaded Leonardo's music tracks for the Forest and the Beach.